feat: #127 알림 상세 내역 조회 Feature 분리
